[SLPVectorizer][NVPTX] Customize getBuildVectorCost for NVPTX
We've observed that the SLPVectorizer is too conservative on NVPTX because it
over-estimates the cost to build a vector. PTX has a single `mov` instruction
that can build <2 x half> vectors from scalars, however the SLPVectorizer
estimates the cost as 2 insert elements.

To fix this I add `TargetTransformInfo::getBuildVectorCost` so the
target can optionally specify the exact cost.